+++ CHANGES.rst
@@ -11,6 +11,8 @@ New Features
 
 - Plots within plugins can now be popped-out into their own windows. [#2254]
 
+- The ``specviz.load_spectrum`` method is deprecated; use ``specviz.load_data`` instead. [#2273]
+
 Cubeviz
 ^^^^^^^